+++ HISTORY.md
@@ -1,5 +1,12 @@
 # History
 
+## not yet released, version 2.0.0-SNAPSHOT
+
+*Breaking changes:*
+
+- Fixed binary prefix and SI prefix incorrectly used for byte. Now following
+  SI standards (`1 KiB == 1024 B`, `1 kB == 1000 B`).
+
 
 ## not yet released, version 1.2.1-SNAPSHOT